cpufreq: zzmoove: A Major fix for some bugs
defined wake hooks in the code so that the governor works perfectly.
lack of this definition leads to screen not waking up after locking the phone>
this fixed this too.

Replaced USE_LCD_NOTIFIER to WAKE_HOOKS_DEFINED so that it works perfectly with MTK source code.

Some changes to definitions too. Thanks to @zanezam for this governor.
you can find the governor's code https://github.com/zanezam/cpufreq-governor-zzmoove HERE

Signed-off-by: Swapnil Solanki <swapnil133609@gmail.com>