cpu: Extend cpu core control by having an option to link core4 online… … status to core3
mpdecision keeps core 4 online all of the time, which is a bit a waste of resources.
By implementing a new control option with value 3 for only the 4th core, the core will
go offline but its online status will be mirrored with the status of the 3rd core.
In a way, core 4 is locked to core 3 by that option.